requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langchain==0.0.350
langchain-openai==0.0.2
langchain-core==0.1.0
requests==2.31.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
gitpython==3.1.40
packaging==23.2
toml==0.10.2
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
aiohttp==3.9.1
python-dotenv==1.0.0
